# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ETGA API Listing",
        "description": "This is the List of APIs that should be used across ETGA and Tourguides ( Dashboard ETGA, Dashboard Tourguides, Website ETGA, Website Tourguides.me)",
        "contact": {
            "email": "nada@technologies.ae"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/availabletimings": {
            "get": {
                "tags": [
                    "AvailableTimings"
                ],
                "summary": "Get list of available timings",
                "operationId": "getAvailableTimings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AvailableTiming"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "AvailableTimings"
                ],
                "summary": "Create a new available timing",
                "operationId": "createAvailableTiming",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AvailableTiming"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Available timing created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailableTiming"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/availabletimings/{id}": {
            "get": {
                "tags": [
                    "AvailableTimings"
                ],
                "summary": "Get available timing by ID",
                "operationId": "getAvailableTimingById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailableTiming"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "AvailableTimings"
                ],
                "summary": "Update available timing",
                "operationId": "updateAvailableTiming",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AvailableTiming"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailableTiming"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AvailableTimings"
                ],
                "summary": "Delete available timing",
                "operationId": "deleteAvailableTiming",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AvailableTiming": {
                "title": "Available Timing",
                "required": [
                    "tour_guide_id",
                    "available_from",
                    "available_to"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the available timing",
                        "type": "integer"
                    },
                    "tour_guide_id": {
                        "description": "ID of the tour guide",
                        "type": "integer"
                    },
                    "available_from": {
                        "description": "Start time of availability",
                        "type": "string",
                        "format": "date-time"
                    },
                    "available_to": {
                        "description": "End time of availability",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}